pyproject.toml
[tool.poetry]
name = "zaida"
version = "0.3.0"
authors = [
  "Mihail Costin Leonte <mail@mcleonte.com>"
]
readme = "README.md"
description = "Zaida AI assistant"
repository = "https://github.com/mcleonte/zaida.ai"

[tool.poetry.dependencies]
python = "^3.8.1,<3.11"
sounddevice = "^0.4.5"
websockets = "^10.4"
SpeechRecognition = "^3.10.0"
PyAudio = "^0.2.13"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
debugpy = "^1.6.4"
python-lsp-server = {extras = ["yapf", "pylint"], version = "^1.7.0"}
pynvim = "^0.5.0"
langchain = {extras = ["all"], version = "^0.1.10"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

[tool.poetry.scripts]
zaida = "zaida:main"
